convert to new plugin2 style
This means that this plugin is no longer compatible with Dancer versions
<= 0.166001. Dancer2 dependency will be updated once the next release
version is known and then this plugin should be released to CPAN
immediately after then next Dancer2 release.

This change is needed since this plugin is used by another plugin
(Dancer2::Plugin::Auth::Extensible::Provider::Usergroup) and so must be
fully compliant with plugin2 before this other plugin can work with
plugin2.